rename State to RootState and appReducers to rootReducers
