+++ src/elements/session-element.ts
@@ -5,7 +5,7 @@ import { ReduxMixin } from '../mixins/redux-mixin';
 import { openDialog } from '../store/dialogs/actions';
 import { DIALOGS } from '../store/dialogs/types';
 import { setUserFeaturedSessions } from '../store/sessions/actions';
-import { store } from '../store/store';
+import { store } from '../store';
 import { showToast } from '../store/toast/actions';
 import { TempAny } from '../temp-any';
 import { getVariableColor, toggleQueryParam } from '../utils/functions';
