+++ src/elements/subscribe-form-footer.ts
@@ -2,7 +2,7 @@ import '@polymer/paper-input/paper-input';
 import { PaperInputElement } from '@polymer/paper-input/paper-input';
 import { html, PolymerElement } from '@polymer/polymer';
 import { ReduxMixin } from '../mixins/redux-mixin';
-import { State, store } from '../store/store';
+import { RootState, store } from '../store';
 import { subscribe } from '../store/subscribe/actions';
 import './hoverboard-icons';
 import './shared-styles';
@@ -106,7 +106,7 @@ class SubscribeFormFooter extends ReduxMixin(PolymerElement) {
     };
   }
 
-  stateChanged(state: State) {
+  stateChanged(state: RootState) {
     this.setProperties({
       subscribed: state.subscribed,
     });
