+++ src/elements/toast-element.ts
@@ -1,7 +1,7 @@
 import '@polymer/paper-toast/paper-toast';
 import { html, PolymerElement } from '@polymer/polymer';
 import { ReduxMixin } from '../mixins/redux-mixin';
-import { State } from '../store/store';
+import { RootState } from '../store';
 import { hideToast } from '../store/toast/actions';
 import { TempAny } from '../temp-any';
 import './shared-styles';
@@ -73,7 +73,7 @@ class ToastElement extends ReduxMixin(PolymerElement) {
     };
   }
 
-  stateChanged(state: State) {
+  stateChanged(state: RootState) {
     this.setProperties({
       toast: state.toast,
       viewport: state.ui.viewport,
