+++ src/mixins/redux-mixin.ts
@@ -1,4 +1,4 @@
 import { connect } from 'pwa-helpers';
-import { store } from '../store/store';
+import { store } from '../store';
 
 export const ReduxMixin = connect(store);
